Remove other interfaces
